Extract budget file related server handlers from main.ts to server/budgetfiles/app.ts
