+++ packages/loot-core/src/server/spreadsheet/spreadsheet.ts
@@ -3,6 +3,7 @@ import mitt from 'mitt';
 
 import { QueryState } from '../../shared/query';
 import { compileQuery, runCompiledQuery, schema, schemaConfig } from '../aql';
+import { BudgetType } from '../prefs';
 
 import { Graph } from './graph-data-structure';
 import { unresolveName, resolveName } from './util';
@@ -20,7 +21,10 @@ export type Node = {
 };
 
 export class Spreadsheet {
-  _meta;
+  _meta: {
+    createdMonths: Set<string>;
+    budgetType: BudgetType;
+  };
   cacheBarrier;
   computeQueue;
   dirtyCells;
@@ -44,6 +48,7 @@ export class Spreadsheet {
     this.events = mitt();
     this._meta = {
       createdMonths: new Set(),
+      budgetType: 'rollover',
     };
   }
