+++ packages/loot-core/src/types/server-handlers.d.ts
@@ -1,11 +1,8 @@
 import { Backup } from '../server/backups';
-import { RemoteFile } from '../server/cloud-storage';
 import { Message } from '../server/sync';
 import { QueryState } from '../shared/query';
 
-import { Budget } from './budget';
 import { OpenIdConfig } from './models/openid';
-import { EmptyObject } from './util';
 
 export interface ServerHandlers {
   undo: () => Promise<void>;
@@ -103,75 +100,6 @@ export interface ServerHandlers {
     | { error: { message: string; reason: string; meta: unknown } }
     | { messages: Message[] }
   >;
-
-  'validate-budget-name': (arg: {
-    name: string;
-  }) => Promise<{ valid: boolean; message?: string }>;
-
-  'unique-budget-name': (arg: { name: string }) => Promise<string>;
-
-  'get-budgets': () => Promise<Budget[]>;
-
-  'get-remote-files': () => Promise<RemoteFile[]>;
-
-  'get-user-file-info': (fileId: string) => Promise<RemoteFile | null>;
-
-  'reset-budget-cache': () => Promise<unknown>;
-
-  'upload-budget': (arg: { id }) => Promise<{ error?: string }>;
-
-  'download-budget': (arg: { fileId; replace? }) => Promise<{ error; id }>;
-
-  'sync-budget': () => Promise<{
-    error?: { message: string; reason: string; meta: unknown };
-  }>;
-
-  'load-budget': (arg: { id: string }) => Promise<{ error }>;
-
-  'create-demo-budget': () => Promise<unknown>;
-
-  'close-budget': () => Promise<'ok'>;
-
-  'delete-budget': (arg: {
-    id?: string | undefined;
-    cloudFileId?: string | undefined;
-  }) => Promise<'ok' | 'fail'>;
-
-  /**
-   * Duplicates a budget file.
-   * @param {Object} arg - The arguments for duplicating a budget.
-   * @param {string} [arg.id] - The ID of the local budget to duplicate.
-   * @param {string} [arg.cloudId] - The ID of the cloud-synced budget to duplicate.
-   * @param {string} arg.newName - The name for the duplicated budget.
-   * @param {boolean} [arg.cloudSync] - Whether to sync the duplicated budget to the cloud.
-   * @returns {Promise<string>} The ID of the newly created budget.
-   */
-  'duplicate-budget': (arg: {
-    id?: string | undefined;
-    cloudId?: string | undefined;
-    newName: string;
-    cloudSync?: boolean;
-    open: 'none' | 'original' | 'copy';
-  }) => Promise<string>;
-
-  'create-budget': (arg: {
-    budgetName?;
-    avoidUpload?;
-    testMode?: boolean;
-    testBudgetId?;
-  }) => Promise<unknown>;
-
-  'import-budget': (arg: {
-    filepath: string;
-    type: 'ynab4' | 'ynab5' | 'actual';
-  }) => Promise<{ error?: string }>;
-
-  'export-budget': () => Promise<{ data: Buffer } | { error: string }>;
-
-  'upload-file-web': (arg: {
-    filename: string;
-    contents: ArrayBuffer;
-  }) => Promise<EmptyObject | null>;
 
   'backups-get': (arg: { id: string }) => Promise<Backup[]>;
